+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Not working update of replicaset roles
 - Not working update of container env vars
+- Problem with a non-existent leader of cluster
 
 ## [0.0.8] - 2020-12-16